sim.src,support,test: replace uint with unsigned/size_t
`uint` and friends are only for backward compatibility. they are not
standardized. and is only available in glibc when `_GNU_SOURCE` is defined.
other implementations of libc might or might not define this type.

we should avoid using it. `unsigned` is used to replace `uint`. and
as an alternative, `size_t` is used when `uint` is used as an index of
array, as `std::size_t` is the index's type used by
`std::vector::operator[]`.

see also https://gcc.gnu.org/bugzilla/show_bug.cgi?id=59945

+++ test/test_dmclock_server.cc
@@ -812,7 +812,7 @@ namespace crimson {
       info2.push_back(dmc::ClientInfo(0.0, 200.0, 0.0));
       info2.push_back(dmc::ClientInfo(0.0, 50.0, 0.0));
 
-      uint cli_info_group = 0;
+      size_t cli_info_group = 0;
 
       QueueRef pq;
